Port inline_subschemas tests to new integration test style
