+++ schemars/tests/integration/main.rs
@@ -26,6 +26,7 @@ mod from_value;
 mod garde;
 #[cfg(feature = "indexmap2")]
 mod indexmap;
+mod inline_subschemas;
 mod std_types;
 
 mod prelude {
