+++ schemars/tests/integration/snapshots/schemars/tests/integration/inline_subschemas.rs~struct_normal.de.json
