+++ schemars/tests/integration/snapshots/schemars/tests/integration/inline_subschemas.rs~struct_normal.ser.json
@@ -1,0 +1,23 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "title": "MyJob",
+  "type": "object",
+  "properties": {
+    "spec": {
+      "type": "object",
+      "properties": {
+        "replicas": {
+          "type": "integer",
+          "format": "uint32",
+          "minimum": 0
+        }
+      },
+      "required": [
+        "replicas"
+      ]
+    }
+  },
+  "required": [
+    "spec"
+  ]
+}
