+++ schemars/tests/integration/snapshots/schemars/tests/integration/inline_subschemas.rs~struct_recursive.ser.json
@@ -1,0 +1,70 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "title": "RecursiveOuter",
+  "type": "object",
+  "properties": {
+    "direct": {
+      "anyOf": [
+        {
+          "$ref": "#/$defs/RecursiveOuter"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
+    "indirect": {
+      "type": [
+        "object",
+        "null"
+      ],
+      "properties": {
+        "recursive": {
+          "$ref": "#/$defs/RecursiveOuter"
+        }
+      },
+      "required": [
+        "recursive"
+      ]
+    }
+  },
+  "required": [
+    "direct",
+    "indirect"
+  ],
+  "$defs": {
+    "RecursiveOuter": {
+      "type": "object",
+      "properties": {
+        "direct": {
+          "anyOf": [
+            {
+              "$ref": "#/$defs/RecursiveOuter"
+            },
+            {
+              "type": "null"
+            }
+          ]
+        },
+        "indirect": {
+          "type": [
+            "object",
+            "null"
+          ],
+          "properties": {
+            "recursive": {
+              "$ref": "#/$defs/RecursiveOuter"
+            }
+          },
+          "required": [
+            "recursive"
+          ]
+        }
+      },
+      "required": [
+        "direct",
+        "indirect"
+      ]
+    }
+  }
+}
